requirements.txt
# -----------------------------------------------------------------------------
# 卵巢 PNG→TIFF 管线（scripts/ov_processing_gpu.py、ov_processing_new_monitor*.py、
# ov_png_folder_to_tif_v2.py）在写出阶段使用 pyvips。
# pip 安装 pyvips 后，运行环境仍需系统级 libvips（例如 RHEL/EL: dnf install vips-devel；
# 或使用 conda-forge: conda install -c conda-forge libvips pyvips）。
# -----------------------------------------------------------------------------
absl-py==2.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
autograd==1.8.0
autograd-gamma==0.5.0
braceexpand==0.1.7
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.2
cuda-bindings==12.9.4
cuda-pathfinder==1.4.0
cycler==0.12.1
ecos==2.0.14
einops==0.8.2
fairscale==0.4.13
filelock==3.25.0
fonttools==4.61.1
formulaic==1.2.1
fsspec==2026.2.0
fvcore==0.1.5.post20221221
gitdb==4.0.12
GitPython==3.1.46
grpcio==1.78.0
h5py==3.15.1
huggingface-hub==0.20.2
idna==3.11
ImageIO==2.37.3
interface-meta==1.3.0
iopath==0.1.10
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
lazy-loader==0.5
lifelines==0.30.2
Markdown==3.10.2
MarkupSafe==3.0.3
matplotlib==3.10.8
monai==1.5.2
mpmath==1.3.0
narwhals==2.17.0
networkx==3.4.2
ninja==1.11.1.1
numexpr==2.14.1
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==13.590.48
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
opencv-python-headless==4.13.0.92
openslide-bin==4.0.0.12
openslide-python==1.4.3
osqp==0.6.7.post3
packaging==23.2
pandas==2.3.3
pillow==12.1.1
platformdirs==4.9.2
portalocker==3.2.0
protobuf==6.33.5
psutil==7.2.2
pydantic==2.12.5
pydantic_core==2.41.5
pynvml==13.0.1
pyparsing==3.3.2
pyvips==3.1.1
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.3
qdldl==0.1.9.post1
regex==2026.2.28
requests==2.32.5
safetensors==0.7.0
scikit-image==0.25.2
scikit-learn==1.7.2
scikit-survival==0.25.0
scipy==1.15.3
seaborn==0.13.2
sentry-sdk==2.54.0
six==1.17.0
smmap==5.0.2
sympy==1.14.0
tabulate==0.10.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.3.0
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==1.0.25
tokenizers==0.15.2
torch==2.10.0
torchmetrics==0.10.3
torchvision==0.25.0
tqdm==4.67.3
transformers==4.36.2
triton==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wandb==0.25.0
webdataset==1.0.2
Werkzeug==3.1.6
wrapt==2.1.1
xformers==0.0.35
yacs==0.1.8
PySocks>=1.7.1
